Update run.bat
No longer installs every single time

+++ run.bat
@@ -1,7 +1,3 @@
-::::::::::::::::::::::::::::::::::::::::::::
-:: Automatically check & get admin rights ::
-::::::::::::::::::::::::::::::::::::::::::::
-
 @echo off
 CLS
 ECHO.
@@ -23,7 +19,7 @@ if '%1'=='ELEV' (echo ELEV & shift /1 & goto gotPrivileges)
 ECHO Set UAC = CreateObject^("Shell.Application"^) > "%vbsGetPrivileges%"
 ECHO args = "ELEV " >> "%vbsGetPrivileges%"
 ECHO For Each strArg in WScript.Arguments >> "%vbsGetPrivileges%"
-ECHO args = args ^& strArg ^& " "  >> "%vbsGetPrivileges%"
+ECHO args = args ^& strArg ^& " " >> "%vbsGetPrivileges%"
 ECHO Next >> "%vbsGetPrivileges%"
 ECHO UAC.ShellExecute "!batchPath!", args, "", "runas", 1 >> "%vbsGetPrivileges%"
 "%SystemRoot%\System32\WScript.exe" "%vbsGetPrivileges%" %*
@@ -32,30 +28,32 @@ exit /B
 :gotPrivileges
 setlocal & pushd .
 cd /d %~dp0
-if '%1'=='ELEV' (del "%vbsGetPrivileges%" 1>nul 2>nul  &  shift /1)
+if '%1'=='ELEV' (del "%vbsGetPrivileges%" 1>nul 2>nul & shift /1)
 
 title Setup
 set PATH=%PATH%;%~dp0
 
-echo Installing...
-python -m venv bot-env
-call .\bot-env\Scripts\activate.bat
-pip install curl_cffi
-pip install fake_useragent
-pip install httpx
-pip install asyncio
-pip install python-dotenv
-pip install pyYAML
-pip install requests
-pip install flask
-pip install bardapi
+if not exist bot-env (
+    echo 'bot-env' folder not found. Installing...
+    python -m venv bot-env
+    call .\bot-env\Scripts\activate.bat
+    pip install curl_cffi
+    pip install fake_useragent
+    pip install httpx
+    pip install asyncio
+    pip install python-dotenv
+    pip install pyYAML
+    pip install requests
+    pip install flask
+    pip install bardapi
 
-pip install -U discord.py-self
-
-echo Installed.
+    pip install -U discord.py-self
+    echo Installed.
+) else (
+    call .\bot-env\Scripts\activate.bat
+)
 
 echo Starting bot...
- 
 python main.py
 pause
 cmd /k
